backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "replicate>=1.0.7",
    "uvicorn>=0.46.0",
]
